chore: remove v2 ref after endpoint refactor
