+++ bc_obps/registration/api/v2/transfer_events.py
@@ -1,6 +1,6 @@
 from typing import List, Literal, Optional, Tuple
 from registration.models.event.transfer_event import TransferEvent
-from registration.schema.v1.transfer_event import TransferEventFilterSchema, TransferEventListOut
+from registration.schema.v2.transfer_event import TransferEventFilterSchema, TransferEventListOut
 from service.transfer_event_service import TransferEventService
 from common.permissions import authorize
 from django.http import HttpRequest
@@ -15,6 +15,7 @@ from registration.schema.generic import Message
 from django.db.models import QuerySet
 from common.api.utils import get_current_user_guid
 from registration.schema.v2.transfer_event import TransferEventCreateIn, TransferEventOut
+
 
 @router.get(
     "/transfer-events",
@@ -35,6 +36,7 @@ def list_transfer_events(
 ) -> QuerySet[TransferEvent]:
     # NOTE: PageNumberPagination raises an error if we pass the response as a tuple (like 200, ...)
     return TransferEventService.list_transfer_events(sort_field, sort_order, filters)
+
 
 @router.post(
     "/transfer-events",
