+++ bciers/libs/actions/src/api/getOperationsByOperatorId.ts
@@ -2,7 +2,7 @@ import { actionHandler } from "@bciers/actions";
 
 async function getOperationsByOperatorId(operatorId: string) {
   return actionHandler(
-    `registration/v2/operators/${operatorId}/operations`,
+    `registration/operators/${operatorId}/operations`,
     "GET",
     "",
   );
